Basic progressive addition of cells working:
- tested with two adjoining cells
- little consideration of performance
- basic mesh generation on demand

+++ Assets/Code/Voronoi/VoronoiUtil.cs
@@ -114,7 +114,7 @@ namespace Growth.Voronoi
             return v;
         }
 
-        public static IProgressiveVoronoi CreateProgressiveVoronoi(float size, float tolerance, ClRand random)
+        public static IProgressiveVoronoi CreateProgressiveVoronoi(int size, float tolerance, ClRand random)
         {
             return new ProgressiveVoronoi(size, tolerance, random);
         }
